Created BanDao, HibernateBanDao; Created BanTypeDao, HibernateBanTypeDao;

+++ src/main/java/dao/banType/BanTypeDao.java
@@ -3,5 +3,5 @@ package dao.banType;
 import dao.CrudDao;
 import entity.BanType;
 
-public interface BanTypeDao extends CrudDao<Long, BanType> {
+public interface BanTypeDao extends CrudDao<Integer, BanType> {
 }
